requirements.txt
amqp
anyjson
appdirs
appnope
asn1crypto
astroid
Babel
backports-abc
backports.functools-lru-cache
backports.shutil-get-terminal-size
backports.ssl-match-hostname
beautifulsoup4
billiard
boto
cached-property
celery
certifi
cffi
click
configparser
coverage
cryptography
decorator
defusedxml
diff-match-patch
Django
django-annoying
django-appconf
django-autocomplete-light
django-braces
django-celery
django-compressor
django-debug-toolbar
django-extensions
-e git+https://github.com/ProminentEdge/django-favit.git@1eb9b1dbbfb65667695da08b3f16c328d1ee74c4#egg=django_favit
django-fixture-magic
-e git+https://github.com/meilinger/django-generic-m2m@1bcf600f2b40a1b56d211b00a01d688553a8be4f#egg=django_generic_m2m
django-guardian
django-import-export
django-invitations
django-map-widgets
django-nose
-e git://github.com/ProminentEdge/django-osgeo-importer.git@45aa4fb1ee091416c761a9906c457014c1a7251c#egg=django_osgeo_importer
django-phonenumber-field
django-queryset-csv
django-registration-redux
django-reversion
django-slack
django-static-sitemaps
django-statsd-mozilla
django-storages
django-tastypie
django-widget-tweaks
enum34
et-xmlfile
-e git+https://github.com/FireCARES/fire-risk.git@51be2749f4f104e5111ee0263932236cc371b562#egg=fire_risk
flake8
flower
funcsigs
futures
geopy
gunicorn
idna
ipaddress
ipdb
ipython
ipython-genutils
isodate
isort
jdcal
jellyfish
jsonfield
kombu
lazy-object-proxy
lxml
matplotlib
mccabe
mock
ndg-httpsclient
nose
numpy
oauthlib
odfpy
openpyxl
packaging
pandas
pathlib2
pbr
pep8
pexpect
phonenumbers
pickleshare
Pillow
prompt-toolkit
psycopg2
ptvsd
ptyprocess
py
pyasn1
pycparser
pyflakes
Pygments
PyJWT
pylibmc
pylint
pyOpenSSL
pyparsing
pytest
pytest-cov
python-dateutil
python-memcached
python-mimeparse
pytz
PyYAML
requests
requests-mock
requests-oauthlib
scandir
scipy
simplegeneric
singledispatch
six
sqlparse
statsd
tablib
tornado
traitlets
unicodecsv
us
wcwidth
wrapt
xlrd
xlwt
zeep
